Move ColorResolver registration to its own manager class
