+++ patches/minecraft/net/minecraft/client/Minecraft.java.patch
@@ -81,7 +81,7 @@
 +      net.minecraftforge.client.gui.overlay.GuiOverlayManager.init();
 +      net.minecraftforge.client.DimensionSpecialEffectsManager.init();
 +      net.minecraftforge.client.NamedRenderTypeManager.init();
-+      net.minecraftforge.client.ForgeHooksClient.registerColorResolvers();
++      net.minecraftforge.client.ColorResolverManager.registerColorResolvers();
 +
        this.f_90990_.m_85409_(this.f_91066_.m_231817_().m_231551_());
        this.f_90990_.m_85424_(this.f_91066_.m_232123_().m_231551_());
